# Fix locks in schema manager