--- a/adapters/repos/db/shard.go
+++ b/adapters/repos/db/shard.go
@@ -86,7 +86,7 @@ type job struct {
 }
 
 func NewShard(ctx context.Context, promMetrics *monitoring.PrometheusMetrics,
-	shardName string, index *Index,
+	shardName string, index *Index, class *models.Class,
 ) (*Shard, error) {
 	before := time.Now()
 
@@ -134,7 +134,7 @@ func NewShard(ctx context.Context, promMetrics *monitoring.PrometheusMetrics,
 		defer s.vectorIndex.PostStartup()
 	}
 
-	if err := s.initNonVector(ctx); err != nil {
+	if err := s.initNonVector(ctx, class); err != nil {
 		return nil, errors.Wrapf(err, "init shard %q", s.ID())
 	}
 
@@ -197,7 +197,7 @@ func (s *Shard) initVectorIndex(
 	return nil
 }
 
-func (s *Shard) initNonVector(ctx context.Context) error {
+func (s *Shard) initNonVector(ctx context.Context, class *models.Class) error {
 	err := s.initDBFile(ctx)
 	if err != nil {
 		return errors.Wrapf(err, "init shard %q: shard db", s.ID())
@@ -224,7 +224,7 @@ func (s *Shard) initNonVector(ctx context.Context) error {
 	}
 	s.propLengths = propLengths
 
-	if err := s.initProperties(); err != nil {
+	if err := s.initProperties(class); err != nil {
 		return errors.Wrapf(err, "init shard %q: init per property indices", s.ID())
 	}
 
@@ -406,6 +406,14 @@ func (s *Shard) addTimestampProperties(ctx context.Context) error {
 func (s *Shard) addPropertyLength(ctx context.Context, prop *models.Property) error {
 	if s.isReadOnly() {
 		return storagestate.ErrStatusReadOnly
+	}
+	dt := schema.DataType(prop.DataType[0])
+	// some datatypes are not added to the inverted index, so we can skip them here
+	switch dt {
+	case schema.DataTypeGeoCoordinates, schema.DataTypePhoneNumber, schema.DataTypeBlob, schema.DataTypeInt,
+		schema.DataTypeNumber, schema.DataTypeBoolean, schema.DataTypeDate:
+		return nil
+	default:
 	}
 
 	err := s.store.CreateOrLoadBucket(ctx,