--- a/entities/deepcopy/models_deepcopy.go
+++ b/entities/deepcopy/models_deepcopy.go
@@ -1,0 +1,96 @@
+//                           _       _
+// __      _____  __ ___   ___  __ _| |_ ___
+// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
+//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
+//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
+//
+//  Copyright © 2016 - 2022 SeMI Technologies B.V. All rights reserved.
+//
+//  CONTACT: hello@semi.technology
+//
+
+package deepcopy
+
+import "github.com/semi-technologies/weaviate/entities/models"
+
+func Schema(s *models.Schema) *models.Schema {
+	classes := make([]*models.Class, len(s.Classes))
+	for i, class := range s.Classes {
+		classes[i] = Class(class)
+	}
+
+	return &models.Schema{Name: s.Name, Maintainer: s.Maintainer, Classes: classes}
+}
+
+func Class(c *models.Class) *models.Class {
+	if c == nil {
+		return nil
+	}
+
+	var properties []*models.Property = nil
+	if c.Properties != nil {
+		properties = make([]*models.Property, len(c.Properties))
+		for i, prop := range c.Properties {
+			properties[i] = Prop(prop)
+		}
+	}
+	var ReplicationConf *models.ReplicationConfig = nil
+	if c.ReplicationConfig != nil {
+		ReplicationConf = &models.ReplicationConfig{Factor: c.ReplicationConfig.Factor}
+	}
+
+	return &models.Class{
+		Class:               c.Class,
+		Description:         c.Description,
+		ModuleConfig:        c.ModuleConfig,
+		ShardingConfig:      c.ShardingConfig,
+		VectorIndexConfig:   c.VectorIndexConfig,
+		VectorIndexType:     c.VectorIndexType,
+		ReplicationConfig:   ReplicationConf,
+		Vectorizer:          c.Vectorizer,
+		InvertedIndexConfig: InvertedIndexConfig(c.InvertedIndexConfig),
+		Properties:          properties,
+	}
+}
+
+func Prop(p *models.Property) *models.Property {
+	var IndexInverted *bool = nil
+	if p.IndexInverted != nil {
+		b := *(p.IndexInverted)
+		IndexInverted = &b
+	}
+
+	return &models.Property{
+		DataType:      p.DataType,
+		Description:   p.Description,
+		ModuleConfig:  p.ModuleConfig,
+		Name:          p.Name,
+		Tokenization:  p.Tokenization,
+		IndexInverted: IndexInverted,
+	}
+}
+
+func InvertedIndexConfig(i *models.InvertedIndexConfig) *models.InvertedIndexConfig {
+	if i == nil {
+		return nil
+	}
+
+	var Bm25 *models.BM25Config = nil
+	if i.Bm25 != nil {
+		Bm25 = &models.BM25Config{B: i.Bm25.B, K1: i.Bm25.K1}
+	}
+
+	var stopwords *models.StopwordConfig = nil
+	if i.Stopwords != nil {
+		stopwords = &models.StopwordConfig{Additions: i.Stopwords.Additions, Preset: i.Stopwords.Preset, Removals: i.Stopwords.Removals}
+	}
+
+	return &models.InvertedIndexConfig{
+		Bm25:                   Bm25,
+		CleanupIntervalSeconds: i.CleanupIntervalSeconds,
+		IndexNullState:         i.IndexNullState,
+		IndexPropertyLength:    i.IndexPropertyLength,
+		IndexTimestamps:        i.IndexTimestamps,
+		Stopwords:              stopwords,
+	}
+}