--- a/usecases/objects/references_update.go
+++ b/usecases/objects/references_update.go
@@ -70,7 +70,7 @@ func (m *Manager) UpdateObjectReferences(
 	}
 	defer unlock()
 
-	validator := validation.New(schema.Schema{}, m.vectorRepo.Exists, m.config)
+	validator := validation.New(m.vectorRepo.Exists, m.config)
 	if err := input.validate(ctx, principal, validator, m.schemaManager); err != nil {
 		return &Error{"bad inputs", StatusBadRequest, err}
 	}