… adapters/repos/db/helper_for_test.go
… usecases/modules/vectorizer_test.go
… usecases/objects/fakes_for_test.go
… usecases/objects/validation/phone_numbers_test.go
… usecases/objects/validation/properties_validation_test.go
… usecases/schema/add_property.go
… usecases/schema/authorization_test.go
… usecases/schema/delete.go
… usecases/schema/get.go
… usecases/schema/incoming_commit.go
… usecases/schema/manager.go
… usecases/schema/meta.go
… usecases/schema/update.go
… usecases/schema/validation.go