### What's being changed:

Review commit by commit is recommened :)

The SchemaManager manages the schema and processes all updates to it. It can be used in two ways:
- Process updates/additions/... and change the schema accordingly
- Return copies to other packages for their needs
Both should be threadsafe now.

Before it was returning pointers to its schema to other packages resulting in many data races. Moreover it had circular call patterns (SchemaManager.AddClass=>Something Else => SchemaManager.PublicFunction) that resulted in deadlocks when trying to lock the public functions.

Motivation to start this was this [stress test (the last one)](https://gist.github.com/dirkkul/513f065fe9e791246ba77bae3a2b900e) that adds objects with different properties with autoschema. Sadly it still doesn't work (it got better and now sometimes all objects are added and can be queried after) but it can now run without any data races. Moreover the acceptance tests run without any races and deadlocks.

### Performance results

master:
```
Runtime for individual steps:.
Runtime for Delete is 462ms.
Runtime for AddSchema is 6ms.
Runtime for BatchAdd is 16961ms.
Runtime for GetObjects is 8ms.
Runtime for Query is 3900ms.
```


new
```
Runtime for individual steps:.
Runtime for AddSchema is 6ms.
Runtime for BatchAdd is 17217ms.
Runtime for GetObjects is 9ms.
Runtime for Query is 4154ms.
Runtime for Delete is 413ms.
```
### Review checklist


- [x] Chaos pipeline run or not necessary. Link to pipeline: https://github.com/semi-technologies/weaviate-chaos-engineering/actions/runs/3751357243/jobs/6372250298
- [x] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.